backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-generativeai==0.8.3
tavily-python==0.5.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.4